+++ app/Http/Controllers/Backend/FrequencyController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Backend;
 use App\Http\Controllers\Controller;
 use App\Models\Frequency;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class FrequencyController extends Controller
 {
@@ -15,6 +16,7 @@ class FrequencyController extends Controller
      */
     public function index()
     {
+        Gate::authorize('app.frequency.index');
         $frequencies = Frequency::all();
         return view('backend.frequency.index',compact('frequencies'));
     }
@@ -26,6 +28,7 @@ class FrequencyController extends Controller
      */
     public function create()
     {
+        Gate::authorize('app.frequency.create');
         return view('backend.frequency.create');
     }
 
@@ -37,6 +40,7 @@ class FrequencyController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('app.frequency.create');
         $request->validate([
             'name'  => 'required|unique:frequencies'
         ]);
@@ -68,6 +72,7 @@ class FrequencyController extends Controller
      */
     public function edit($id)
     {
+        Gate::authorize('app.frequency.edit');
         $frequency = Frequency::findOrfail($id);
         return view('backend.frequency.create',compact('frequency'));
 
@@ -82,6 +87,7 @@ class FrequencyController extends Controller
      */
     public function update(Request $request, $id)
     {
+        Gate::authorize('app.frequency.edit');
         $request->validate([
             'name'  => 'required'
         ]);
@@ -102,6 +108,7 @@ class FrequencyController extends Controller
      */
     public function destroy($id)
     {
+        Gate::authorize('app.frequency.destroy');
         $frequency = Frequency::findOrfail($id);
         $frequency->delete();
         return response()->json($frequency);
