+++ app/Http/Controllers/Backend/QuantityTypeController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Backend;
 use App\Http\Controllers\Controller;
 use App\Models\QuantityType;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class QuantityTypeController extends Controller
 {
@@ -15,6 +16,7 @@ class QuantityTypeController extends Controller
      */
     public function index()
     {
+        Gate::authorize('app.quantitytype.index');
         $quantityTypes = QuantityType::all();
         return view('backend.quantityType.index',compact('quantityTypes'));
     }
@@ -26,6 +28,7 @@ class QuantityTypeController extends Controller
      */
     public function create()
     {
+        Gate::authorize('app.quantitytype.create');
         return view('backend.quantityType.create');
     }
 
@@ -37,6 +40,7 @@ class QuantityTypeController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('app.quantitytype.create');
         $request->validate([
             'name'  => 'required|unique:quantity_types'
         ]);
@@ -68,6 +72,7 @@ class QuantityTypeController extends Controller
      */
     public function edit($id)
     {
+        Gate::authorize('app.quantitytype.edit');
         $quantityType = QuantityType::findOrfail($id);
         return view('backend.quantityType.create',compact('quantityType'));
     }
@@ -81,6 +86,7 @@ class QuantityTypeController extends Controller
      */
     public function update(Request $request, $id)
     {
+        Gate::authorize('app.quantitytype.edit');
         $request->validate([
             'name'  => 'required'
         ]);
@@ -101,6 +107,7 @@ class QuantityTypeController extends Controller
      */
     public function destroy($id)
     {
+        Gate::authorize('app.quantitytype.destroy');
         $quantityType = QuantityType::findOrfail($id);
         $quantityType->delete();
         return response()->json($quantityType);
